+++ config/apiConfig.js
@@ -1,4 +1,4 @@
 // define api root in this config
 export default {
-  apiRoot: "/services"
+  apiRoot: ""
 }
